core+executor: M6.0.2 — subpath exports + test-helpers consolidation
Adds three new subpath exports to @beevibe/core (./auth, ./services/
task-service, ./test-helpers) consistent with the M5 pattern. Promotes
test-helpers from packages/core/src/adapters/postgres/test-helpers.ts
to packages/core/src/test-helpers.ts so it can be exposed as a root
subpath. Removes the executor's duplicate test-helpers.ts (flagged in
M5 simplify but punted) — executor's worker.test.ts now imports from
@beevibe/core/test-helpers.

- packages/core/package.json: 3 new subpath exports
- packages/core/src/test-helpers.ts: moved from adapters/postgres/,
  imports updated, .env path fixed (3 levels up to repo root for both
  source + dist layouts since src/ and dist/ are siblings)
- packages/core/tsconfig.json: drop test-helpers.ts from exclude (it
  now needs to build into dist for the subpath); test-fakes.ts (vitest-
  using auth fakes) added to exclude
- 7 core postgres tests + 1 auth integration test: relative imports
  updated to ../../test-helpers.js / ../test-helpers.js
- packages/executor/src/test-helpers.ts: deleted
- packages/executor/src/worker.test.ts: import from
  @beevibe/core/test-helpers
- packages/executor/tsconfig.json: drop test-helpers.ts from exclude

Build + typecheck + lint green. All 289 unit tests pass except 2 pre-
existing Anthropic LLM live-API tests (credit balance issue, unrelated).
Executor's 25 worker tests pass against the subpath import.
