+++ packages/core/src/test-helpers.ts
@@ -1,13 +1,15 @@
 import { config as loadEnv } from "dotenv";
 import { fileURLToPath } from "node:url";
 import { resolve, dirname } from "node:path";
-import type { Pool } from "./client.js";
-import { createPool } from "./client.js";
+import type { Pool } from "./adapters/postgres/client.js";
+import { createPool } from "./adapters/postgres/client.js";
 
 // Resolve the repo-root .env file so tests work regardless of cwd
 // (vitest runs from the package dir; .env is at the repo root).
+// Both source (packages/core/src/test-helpers.ts) and built dist
+// (packages/core/dist/test-helpers.js) are 3 levels up from repo root.
 const here = dirname(fileURLToPath(import.meta.url));
-const repoRootEnv = resolve(here, "../../../../../.env");
+const repoRootEnv = resolve(here, "../../../.env");
 loadEnv({ path: repoRootEnv });
 
 const ALL_TABLES = [
