+++ packages/core/tsconfig.json
@@ -5,5 +5,5 @@
     "rootDir": "./src"
   },
   "include": ["src/**/*"],
-  "exclude": ["dist", "node_modules", "**/*.test.ts", "**/test-helpers.ts"]
+  "exclude": ["dist", "node_modules", "**/*.test.ts", "**/test-fakes.ts"]
 }
